+++ R/regression_frame_glmmTMB.R
@@ -52,6 +52,7 @@ as_regression_frame.glmmTMB <- function(fit,
 
   coefs <- .glmmTMB_coefs(fit, ci_level = ci_level)
   coefs <- .attach_ame_to_frame_coefs(coefs, fit, ci_level, show_columns)
+  coefs <- .attach_partial_chi2_to_frame_coefs(coefs, fit, show_columns)
   info  <- .glmmTMB_info(fit,
                          vcov_kind  = vcov,
                          vcov_label = vcov_label,