+++ R/regression_frame_merMod.R
@@ -46,6 +46,7 @@ as_regression_frame.lmerMod <- function(fit,
 
   coefs <- .merMod_coefs(fit, ci_level = ci_level, family_z = FALSE)
   coefs <- .attach_ame_to_frame_coefs(coefs, fit, ci_level, show_columns)
+  coefs <- .attach_partial_chi2_to_frame_coefs(coefs, fit, show_columns)
   info  <- .merMod_info(fit,
                         vcov_kind  = vcov,
                         vcov_label = vcov_label,
@@ -105,6 +106,7 @@ as_regression_frame.glmerMod <- function(fit,
 
   coefs <- .merMod_coefs(fit, ci_level = ci_level, family_z = TRUE)
   coefs <- .attach_ame_to_frame_coefs(coefs, fit, ci_level, show_columns)
+  coefs <- .attach_partial_chi2_to_frame_coefs(coefs, fit, show_columns)
   info  <- .merMod_info(fit,
                         vcov_kind  = vcov,
                         vcov_label = vcov_label,