+++ R/regression_frame_nlme.R
@@ -44,6 +44,7 @@ as_regression_frame.lme <- function(fit,
 
   coefs <- .lme_coefs(fit, ci_level = ci_level)
   coefs <- .attach_ame_to_frame_coefs(coefs, fit, ci_level, show_columns)
+  coefs <- .attach_partial_chi2_to_frame_coefs(coefs, fit, show_columns)
   info  <- .lme_info(fit,
                      vcov_kind  = vcov,
                      vcov_label = vcov_label,